+++ dynamicbeat/pkg/assets/spaces/spaces.go
@@ -1,0 +1,11 @@
+package spaces
+
+import (
+	"io"
+
+	"github.com/scorestack/scorestack/dynamicbeat/pkg/assets"
+)
+
+func Scorestack() io.Reader {
+	return assets.Read("spaces/scorestack.json")
+}